Fix ci (#16)
